move constants into utils/ folder and out of app/
