--- src/app/constants.ts
@@ -1,1 +1,0 @@
-export const jsonStoreFileName = "app_store.json"
