+++ src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const jsonStoreFileName = 'app_store.json';
